refactor(providers): extract try_authenticate
